+++ src/js/main.js
@@ -1,3 +1,9 @@
+/* eslint-disable indent */
+if (import.meta.env.MODE !== 'development') {
+  // eslint-disable-next-line prettier/prettier
+  import('vite/modulepreload-polyfill')
+}
+
 import '../scss/style.scss'
 
 document.querySelector('#app').innerHTML = `